refactor(cli): the --var door moves into the host crate
nika-cli crossed the 15k production-line wall with the test verb's
cases. The parse and the origin law of --var read only the schema and
the runtime, and the golden test now binds a case through the same
door, so the door lives in nika-cli-host (var_inputs); nika-cli keeps
the env-refusal wrap in run's voice.
